# perf(genAdds): refactor recursive procedure to iterative