--- a/packages/rrweb/src/record/mutation.ts
+++ b/packages/rrweb/src/record/mutation.ts
@@ -739,38 +739,64 @@ export default class MutationBuffer {
   /**
    * Make sure you check if `n`'s parent is blocked before calling this function
    * */
-  private genAdds = (n: Node, target?: Node) => {
-    // this node was already recorded in other buffer, ignore it
-    if (this.processedNodeManager.inOtherBuffer(n, this)) return;
+  private genAddsQueue: [Node, Node | undefined][] = new Array<
+    [Node, Node | undefined]
+  >(1000);
+  private genAdds = (node: Node, t?: Node) => {
+    let rp = -1;
+    let wp = -1;
+    this.genAddsQueue[++wp] = [node, t];
 
-    // if n is added to set, there is no need to travel it and its' children again
-    if (this.addedSet.has(n) || this.movedSet.has(n)) return;
+    while (rp < wp) {
+      const next = this.genAddsQueue[++rp];
+      if (!next) {
+        throw new Error(
+          'Add queue is corrupt, there is no next item to process',
+        );
+      }
+      const [n, target] = next;
 
-    if (this.mirror.hasNode(n)) {
-      if (isIgnored(n, this.mirror, this.slimDOMOptions)) {
+      // this node was already recorded in other buffer, ignore it
+      if (this.processedNodeManager.inOtherBuffer(n, this)) continue;
+
+      // if n is added to set, there is no need to travel it and its' children again
+      if (this.addedSet.has(n) || this.movedSet.has(n)) continue;
+
+      if (this.mirror.hasNode(n)) {
+        if (isIgnored(n, this.mirror, this.slimDOMOptions)) {
+          continue;
+        }
+        this.movedSet.add(n);
+        let targetId: number | null = null;
+        if (target && this.mirror.hasNode(target)) {
+          targetId = this.mirror.getId(target);
+        }
+        if (targetId && targetId !== -1) {
+          this.movedMap[moveKey(this.mirror.getId(n), targetId)] = true;
+        }
+      } else {
+        this.addedSet.add(n);
+        this.droppedSet.delete(n);
+      }
+
+      const isNodeBlocked = isBlocked(
+        n,
+        this.blockClass,
+        this.blockSelector,
+        false,
+      );
+      if (isNodeBlocked) {
         return;
       }
-      this.movedSet.add(n);
-      let targetId: number | null = null;
-      if (target && this.mirror.hasNode(target)) {
-        targetId = this.mirror.getId(target);
-      }
-      if (targetId && targetId !== -1) {
-        this.movedMap[moveKey(this.mirror.getId(n), targetId)] = true;
-      }
-    } else {
-      this.addedSet.add(n);
-      this.droppedSet.delete(n);
-    }
-
-    // if this node is blocked `serializeNode` will turn it into a placeholder element
-    // but we have to remove it's children otherwise they will be added as placeholders too
-    if (!isBlocked(n, this.blockClass, this.blockSelector, false)) {
-      n.childNodes.forEach((childN) => this.genAdds(childN));
+      // if this node is blocked `serializeNode` will turn it into a placeholder element
+      // but we have to remove it's children otherwise they will be added as placeholders too
+      n.childNodes.forEach((childN) => {
+        this.genAddsQueue[++wp] = [childN, undefined];
+      });
       if (hasShadowRoot(n)) {
         n.shadowRoot.childNodes.forEach((childN) => {
           this.processedNodeManager.add(childN, this);
-          this.genAdds(childN, n);
+          this.genAddsQueue[++wp] = [childN, n];
         });
       }
     }