… packages/rrweb/test/__snapshots__/integration.test.ts.snap