Not sure how representative the benchmark is, so I'm adding the screenshot of the profile (deep DOM tree benchmark). Looking at timings, it seems to take ~18ms before and 13ms after my change, or about ~30% gain. Feel free to do more extensive benchmarking and take those timings with a grain of salt as I just ran the benchmarks once

Before
<img width="698" alt="CleanShot 2023-08-08 at 10 11 25@2x" src="https://github.com/rrweb-io/rrweb/assets/9317857/0176935f-ea61-4e9f-b981-fb3c91d96528">

After
<img width="549" alt="CleanShot 2023-08-08 at 10 11 36@2x" src="https://github.com/rrweb-io/rrweb/assets/9317857/b34baec9-ecde-4871-8fc2-d7e106d608f1">

We can see the stack is shallower as expected (and uses less memory, sadly no metrics on that).
A small benefit is that we also remove the overhead of calling into the polymorphic genAdds fn.

The queue has a magic default size of 1000 - feel free to tweak that to some better value, I'm not sure what a sensible default could be here