@eoghanmurray 2023-08-17
We should also include the test changes in this PR also for further review — see the `yarn test:update` command (I can run this for you if you prefer, but you mentioned you had checked the output?)