@JonasBa 2023-08-17
@eoghanmurray no problem, let me make the change.

Is there a reason that you decided to move genAdds inside the `childlist` case here btw?