@eoghanmurray 2023-08-24
> Is there a reason that you decided to move genAdds inside the childlist case here btw?

Do you mean https://github.com/rrweb-io/rrweb/pull/1277/commits/f5f9789270bfed6315f625d89e98b49ffb81ca31 ?
I imagine `genAdds` was originally separate only because it was recursive.