@JonasBa 2023-08-31
I think my dev env must have picked up a different TS version.
What I see in serialize-args is this error
<img width="739" alt="CleanShot 2023-08-30 at 20 51 15@2x" src="https://github.com/rrweb-io/rrweb/assets/9317857/438a223c-fa8c-4c65-be8d-7c846a6003be">

which gets corrected to another error...
<img width="730" alt="CleanShot 2023-08-30 at 20 51 47@2x" src="https://github.com/rrweb-io/rrweb/assets/9317857/d6231897-f8a2-43b7-a04b-4c0c2e1e57c3">

I reverted the changes, it should now be ok.