@eoghanmurray 2023-10-09
Apologies for delay; do you have this running in production?

I've just cherry-picked this PR today and am putting it out in a 'canary' environment to a subset of customers.

Reasons to be cautious: I've fixed a completely separate 'node ordering' issue on the replay side of things, where non-anticipated ordering of nodes in a mutation produced worst case performance and ultimately broke the replay (as it triggered a timeout).
Fix is here https://github.com/rrweb-io/rrweb/pull/1320 so hopefully that acts as a vote in favour of this one as we should be more robust at replay time now.