@JonasBa 2023-10-09
>Apologies for delay; do you have this running in production?
We do not.

>I've just cherry-picked this PR today and am putting it out in a 'canary' environment to a subset of customers.
Perfect, let me know.

I looked at your PR about node ordering issue, but since it's been a while since I looked at this code I lost some context and requirements, so I have a hard time imagining the exact issue you are fixing.