@bruno-garcia 2023-12-01
> I've just cherry-picked this PR today and am putting it out in a 'canary' environment to a subset of customers.

Does it give the confidence we need to merge this one in? Wondering if we should be cherry-picking it too or it merges here and update our fork