@eoghanmurray 2024-04-14
See also #1447 ... I'm trying to work out how these two PRs might interact