@JonasBa 2024-04-14
@eoghanmurray I think my PR already implements the changes from that PR as it performs an in-order traversal and then just proceeds to pop the items from the queue (lifo), which is the essentially the same as iterating backwards as you process items