@eoghanmurray 2024-04-15
Super, I missed that it was LIFO.
No need for `unshift` so