@JonasBa 2024-04-15
Yeah, you want to avoid unshift in this case because it's On. We would have had to use a list in this case. Btw, whats the status of this PR, can we merge this or do you want me to do anything else here?