@eoghanmurray 2024-04-15
I think we should have merged a long time ago, however with something like this it's hard to tell if there isn't some case where it makes things worse (not that I believe this is the case from having examined it).

I meant to put together another 'throw everything at it' mutation benchmark test today with this in mind, but ran into some technical issues.

I think it will be merged soon.