@eoghanmurray 2024-04-15
You could rebase/merge on master which will re-trigger fresh tests; not sure if the indicated test failure is significant (it is no longer available for viewing).