@JonasBa 2024-04-17
@eoghanmurray I think the main change is knowing if the order of insertion/processing matters here? I'm guessing it doesnt as per https://github.com/rrweb-io/rrweb/pull/1398, so we should be safe