@JonasBa 2024-05-28
@eoghanmurray I realized that the benchmarks were so similar because I wasnt rebuilding the packages and running yarn benchmark from within the rrweb package.

I've reran the benchmarks with the following results. The largest change was observed on the append 1000 elements and reverse their order benchmark where perf seems to have improved by 15%. The other benchmarks seem to have all stayed within a few % of the current value. The metrics of each these runs look like standard error, so I dont think we can take them into account

**1121ms -> 1075ms**
'append 70 x 70 x 70 elements' │   3   │ 1121.3333333333333 │ '1128, 1117, 1119'
'append 70 x 70 x 70 elements' │   3   │ 1075.6666666666667 │ '1071, 1094, 1062'

**165ms -> 140ms -15% (largest change)**
'append 1000 elements and reverse their order' │   3   │ 165.33333333333334 │ '165, 168, 163'
'append 1000 elements and reverse their order' │   3   │ 140.33333333333334 │ '139, 141, 141'

**1390ms -> 1401ms**
'real events recorded on bugs.chromium.org' │   3   │   1390   │ '1430, 1348, 1392'
'real events recorded on bugs.chromium.org' │   3   │ 1401.3333333333333 │ '1465, 1388, 1351'

**81.7m -> 82.9ms**
'create 1000x10x2 DOM nodes and remove a bunch of them' │ 'benchmark-dom-mutation-add-and-remove.html' │ 'window.workload()' │  10   │   81.7   │ '84, 77, 81, 92, 76, 77, 95, 80, 79, 76'
'create 1000x10x2 DOM nodes and remove a bunch of them' │ 'benchmark-dom-mutation-add-and-remove.html' │ 'window.workload()' │  10   │   82.9   │ '82, 80, 79, 91, 79, 85, 96, 79, 75, 83'

**54ms -> 56ms**
'create 1000x10 DOM nodes' │ 'benchmark-dom-mutation.html' │ 'window.workload()' │  10   │    54    │ '53, 53, 54, 52, 53, 57, 53, 52, 56, 57'
'create 1000x10 DOM nodes' │ 'benchmark-dom-mutation.html' │ 'window.workload()' │  10   │   56.1   │ '52, 55, 55, 53, 55, 56, 54, 51, 71, 59'

**22.8 -> 21.8**
'create 1000 DOM nodes and append into its previous looped node' │ 'benchmark-dom-mutation-multiple-descendant-add.html' │ 'window.workload()' │   5   │   22.8   │ '23, 23, 22, 23, 23'
'create 1000 DOM nodes and append into its previous looped node' │ 'benchmark-dom-mutation-multiple-descendant-add.html' │ 'window.workload()' │   5   │   21.8   │ '21, 22, 22, 22, 22'

**89.6ms -> 89.8ms**
'create 10000 DOM nodes and move it to new container' │ 'benchmark-dom-mutation-add-and-move.html' │ 'window.workload()' │   5   │   89.6   │ '90, 95, 86, 91, 86'
'create 10000 DOM nodes and move it to new container' │ 'benchmark-dom-mutation-add-and-move.html' │ 'window.workload()' │   5   │   89.8   │ '87, 100, 89, 87, 86' 

**19.3ms -> 17.9ms**
'modify attributes on 10000 DOM nodes' │ 'benchmark-dom-mutation-attributes.html' │ 'window.workload()' │  10   │   19.3   │ '18, 24, 17, 18, 28, 17, 18, 16, 17, 20'
'modify attributes on 10000 DOM nodes' │ 'benchmark-dom-mutation-attributes.html' │ 'window.workload()' │  10   │   17.9   │ '17, 16, 18, 20, 15, 19, 19, 16, 23, 16'